% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freg}
\alias{freg}
\title{(Internal) Defines the regression function used in the simulation by name}
\usage{
freg(x, freg.name)
}
\arguments{
\item{x}{A vector for one-dimensional functions, a matrix for multidimensional ones, containing values for which the regression function has to be computed}

\item{freg.name}{Name of the regression function to use, among "abs", "fried1", "sinus" and "sum"}
}
\value{
A vector of the resulting computed values of the regression function associated to the inputs in \code{x}.
}
\description{
(Internal) Defines the regression function used in the simulation by name
}
