% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horf-package.R
\docType{package}
\name{horf-package}
\alias{horf}
\alias{horf-package}
\title{horf: Implementation of Hold-Out Random Forests and their bias computation}
\description{
This package contains necessary functions to reproduce the simulation experiments of the article Arlot and Genuer (XXXX). In this article the rate of convergence of Hold-Out Random Forests bias is studied. Thus, the package gives an implementation of this Random Forests variant and a plot function to visualize the bias convergence rate of trees v.s. forests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robingenuer/horf}
  \item Report bugs at \url{https://github.com/robingenuer/horf/issues}
}

}
\author{
\strong{Maintainer}: Robin Genuer \email{robin.genuer@u-bordeaux.fr}

Authors:
\itemize{
  \item Sylvain Arlot
}

}
\keyword{internal}
