% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias.R
\name{plot_bias}
\alias{plot_bias}
\title{Plot of the evolution of the bias of forests and trees against the number of leaves of trees, with confidence intervals for each bias estimation and for slopes and intercepts of linear fits to curves}
\usage{
plot_bias(
  forest.comput.bias.res,
  forest.confint.res,
  tree.comput.bias.res,
  tree.confint.res
)
}
\arguments{
\item{forest.comput.bias.res}{Resulting object from \code{comput_bias()} function for forests (with \code{tree = FALSE} in the call of the function)}

\item{forest.confint.res}{Resulting object from \code{comput_confint()} function applied to the content of \code{forest.comput.bias.res}}

\item{tree.comput.bias.res}{Resulting object from \code{comput_bias()} function for trees (with \code{tree = TRUE} in the call of the function)}

\item{tree.confint.res}{Resulting object from \code{comput_confint()} function applied to the content of \code{tree.comput.bias.res}}
}
\value{
Nothing is returned, only a plot is made.
}
\description{
Plot of the evolution of the bias of forests and trees against the number of leaves of trees, with confidence intervals for each bias estimation and for slopes and intercepts of linear fits to curves
}
\examples{
forest_bias_result <- comput_bias(nbleaves = c(2^5, 2^6), freg.name = "sinus",
   xdim = 1, nbobs = 640, nbobs_test = 60, nfor = 10, var_estim = TRUE,
   mc.cores = 2)
forest_confint_result <- comput_confint(forest_bias_result, nbBootstrap = 250)
tree_bias_result <- comput_bias(nbleaves = c(2^5, 2^6), freg.name = "sinus",
   xdim = 1, nbobs = 640, nbobs_test = 60, nfor = 10, tree = TRUE,
   var_estim = TRUE, mc.cores = 2)
tree_confint_result <- comput_confint(tree_bias_result, nbBootstrap = 250)
plot_bias(forest_bias_result, forest_confint_result,
    tree_bias_result, tree_confint_result)
}
