% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simul_data}
\alias{simul_data}
\title{(Internal) Generation of a simulated dataset associated to a given name of the regression function}
\usage{
simul_data(nbobs, xdim, freg.name, sigma = 1/4)
}
\arguments{
\item{nbobs}{The number of observations of the simulated dataset}

\item{xdim}{Dimension of the input space of th regression function}

\item{freg.name}{Name of the regression function to use, among "abs", "fried1", "sinus" and "sum"}

\item{sigma}{The standard deviation of the centered gaussian noise added to computed output values (default to 1/4)}
}
\value{
A list containing the simulated matrix of inputs \code{x} and the simulated one-column matrix of outputs \code{y}.
}
\description{
(Internal) Generation of a simulated dataset associated to a given name of the regression function
}
\examples{
simData <- horandomforest:::simul_data(nbobs = 100, xdim = 1, freg.name = "sinus")
plot(simData)
curve(horandomforest:::freg(x, freg.name = "sinus"), from = 0, to = 1, add = TRUE, col = 2)
}
